Split graphics from sys (region)
